Removed useless and annoying linting git hook
Now you are allow to create commit even if the code doesn't pass linting.

Linting is ensured by the CI which runs linting and you are not allowed to merge a PR
where the CI is failing. On the other hand it is allow to create an push non conform code as
work in progress on a local or personal branch.