Use VECTOR_BUCKET_REGION for s3_connection if this env var is set
Before the migration, the default region for s3 connections was eu-west-1. As the VECTOR_BUCKET has already been migrated
to eu-central-1, the code now checks if a VECTOR_BUCKET_REGION is set. If that is the case, this value will be used
as region for the s3 connection.
This approach is based on the assumption, that S3Connect is only used for the VECTOR_BUCKET and not other buckets, especially
not other buckets in other regions