BGDIINF_SB-2328: introducing a flag INSIDE_DOCKER_IMAGE to Makefile.frankfurt for reducing size of docker image
Fetching the required docker metadata uses git commands, i.e. git
rev-parse, git status, etc., which can only be executed if a .git
directory is present. In order to exclude the .git dir from the docker
image and reduce the image size significantly, a new flas for the
Makefile.frankfurt was introduced: INSIDE_DOCKER_IMAGE with the default
value=False. If the Makefile.frankfurt is invoked from within the Docker
image, it is set to True. With this, the docker metadata is only
generated when necessary and does not cause problems inside the image
with a missing .git dir."