BGDIINF_SB-2426: Set postgres connection timeout
By default connection to postgres don't have a timeout which means that the
application would wait indefinitely to the DB connection if not available.
In production this should never occur, however while developing this can happen
in case of wrong configuration. Therefor set a 30 seconds timeout to have
the application crash with a proper message and backtrace in case of unavailable
DB.