BGDIINF_SB-2425: Fixed local unittests in Makefile.frankfurt
Also include the env file in order to have the variables in the help targets.

Re-include the `test_wmtscapabitlities` test, because they are working. But
still keeps the search tests disabled as they need access to the sphinx host
which we don't have in Frankfurt (the search has been moved in its own services)