BGDIINF_SB-2465: Improved logging and made the application log level configurable
All logs, inclusive the python logs, are delivered by the apache logging system.
Therefore some informations like timestamp, request ID and headers are set
by the apache config.

Also improved the logging format of the apache config with useful information.

Fixed as well the production.ini file. The `use = ...` syntax in PasteDeploy
ini file works only for one section, so the logging configuration has to be
duplicated in both dev.ini.in and prod.ini.in. The `[server:main]` section needs
to defined in both `dev.ini.in` and `prod.ini.in` using the `use` syntax from the
base.