BGDIINF_SB-2475: moved comment in .env.default to the very bottom
When running the test Makefile.frankfurt target on vhost-dev, the name of the VECTOR_BUCKET was not defined.
This was due to a commenting line in .env.default. The bulk export with  in the test
target ignored all entries after the commenting line. Hence all comment lines should be placed below the last
definition of env vars.