BGDIINF_SB-2515: Remove Apache compression
When using the Apache compression delfate module, apache removes the Content-Length
header. Apparently this is due to the chunked encoding feature of apache. I
tried to increase the DeflateBufferSize like mentioned in the docu, but still
the Content-Length for small request was missing.

The main issue is that if the client don't send the "Accept-Encoding" header
CF set it to "Accept-Encoding: gzip" and then decompress the answer. In this
case then the Content-Length header is missing.

For now I removed the compression in service and let CF to do the compression.

In future it might a better idea to do compression directly in the service,
or if we change apache by another wsgi server we might have a better compression.

See BGDIINF_SB-2486 and BGDIINF_SB-2510