Issue BGDIINF_SB-2488 : sdi service crashes on queries with SR IDS 3857 and 4326 Jira ticket : https://jira.swisstopo.ch/browse/BGDIINF_SB-2488
Cause of the problem :

The function aggregating the informations about all zoom levels was
checking through all zoom levels (0 to 28), but the problematic SRIDS
had less than 28 zoom level, encountering a problem within an assertion
which crashed the whole function.

Mitigation :

We use the RESOLUTION table length from the geoadmin tilegrid object
created with the srid, ensuring we no longer can try to access zoom
levels that are not available for a specific srid.