https://jira.swisstopo.ch/browse/BGDIINF_SB-2506
Adapting The Dockerfile and Makefile to use pipenv instead of venv.

removing old requirement file for python 3

moving the pipenv cleaning part to another target, as pipenv --rm crash if pipenv has not been set within codebuild

Reverting some changes, as the goal is to have minimal impact on the
project and how it works.

adding the line which makes the setup for the chsdi project

timestamps are not used yet

We can't remove requirements-py3.txt yet.
Added dev packages category in Pipfile
Correction to the makefile to also install dev packages and use the
setup script when serving

a comment to remember why the file is still here

removing some verbosity during installation

The setup.py script called by pipenv install -e no longer require the
'requirements-py3.txt' file. Instead, it generates a requirements file
on the fly (if it doesn't exist already) to feed the setup function.

This new file is cleaned up by the makefile target "clean" and is
part of the gitignore.