BGDIINF_SB-2554: renamed LOADER_JS_BUCKET_HOST | using Cloudfront-Forwarded-Proto header
LOADER_JS_BUCKET_HOST is now LOADER_JS_BUCKET_LOCALHOST and only used when testing locally. The
bucket host in that case is pointing to dev, instead of the non existend localhost bucket.
For the construction of the URL in loader.js the protocol of CloudFront-Forwarded-Proto header is now used.
If not existing, the request.host_url which inclused the scheme will be used.