BGDIINF_SB-2554: Adapting request globally for all request objects
As not only loader.js endpoint might be affected, it was decided to implement a simple,
global solution that adapta wsgi.url_scheme for all request objects. Hence the
request factory was extended such that wsgi.url_scheme is adapted based on the existence
and value of either CloudFront-Forwarded-Proto or X-Forwarded-Proto headers. This solution
can be deleted, once apache is removed. Afterwards it can be solved in a similar way as in
our flask services that use gunicorn.