Fixed mako with versionned urls
Some makos uses versionned urls, for those we have an apache rewrite url
that should match an official release version optionally followed by a git
hash.

This might be an issue if the git repository contains tag that don't match
our milestone version scheme. Therefore now we check if the GIT_TAG is valid
and if not we use a dummy version for the APP_VERSION.