In a compromise for performances, pipenv install -e . has been removed and is replaced by its predecessor, pip install -e .
The reasoning is the following :
pipenv install -e . append a package to the pipfile, dirtying the git
repository upon call. This could be solved by adding the chsdi package
at the end of the Pipfile, but doing so would crash the docker build if
we tried to install the environment before copying the files.

Thus, we install all standard libraries with pipenv, and call pip3 to
install only the chsdi package at the end, greatly reducing the time
needed for each docker build.