BGDIINF_SB-2486: Moved the new config style url rewrite inside python
For this support I had to move the `add_localizer()` hook from `NewRequest` to
`ContextFound` event. In the first one no routing is done and we cannot take
the `lang` parameter from the path for the new config style urls, while in the
ContextFound the routing has already be done and the `lang` url parameter has
already been parsed.