BGDIINF_SB-2498: Fixed PyProj warning due to update
With the update of pyproj to the latest 3.4 version, a deprecation warning
appeared. The `Proj(init=<authority>:<code>)` construct should not be used
anymore and changed to `CRS` construct. Also with this changes the projection
transformation has been changed by a much more performant method using the new
`Transformer`. Because we have a limited number of CRSs and Transformers those
are now cached in memory in a threading safe manner. Previously the Projections
where also cached in memory using a global variable but this was not thread safe.