Fixing CI manual build and versioning
The docker image was build with the dirty label due to the installation of
kubectl locally inside the repo. This has been fixed by moving kubectl to
~/bin instead of copying.

Manual build on master branch did not pushed the docker image on ECR, because
the `git describe` command returns the tag instead of the ref.