BGDIINF_SB-2950: by default pyramid did not catch all exception and log them
By default pyramid did not logged the exception, they were simply printed on the
console. Except when running locally with waitress server which logged the exception
using the logging configuration.

This made the debugging of crash on PROD harder as we needed to look into the
non json index in kibana and that each backtrace line was treated as a separate
log.

Now with the pyramid_exclog package exception are logged based on the logging
configuration and the exc_info flag is set to true and the backtrace is set
into the field exc_text. This will help to debug crash with kibana.