PB-974: Migrating from TRIAS to OJP 2.0 for open transport data
The API we use to get time table data for open transport stops will be
decomissioned soon. Hence we migrate to OJP 2.0.
The requests to the API and the parsing of the response needed some
adaptions for that. Furthermore, the OJP 2.0 API's response sometimes
has a strange timestamp format, there are sometimes 7 instead of the
expexted 6 digits for the milliseconds. Hence a try except was used to
safely convert to the expected format in that cases.
Also some minor code clean-up was done.