fix: Allow redirects in resource_exists helper function
with the redirect of api3.geo.admin.ch the test has to be adapted.
The resource_exists() function now follows HTTP redirects by setting
allow_redirects=True in the requests.head() call. This fixes the
test_resource_exists test which was failing because api3.geo.admin.ch
returns a 302 redirect instead of 200 OK.